# openapi.json
{"openapi": "3.0.2", "info": {"title": "Coca Cola Inventory Detection", "version": "2023.1.31"}, "paths": {"/healthcheck": {"get": {"summary": "Perform Healthcheck", "description": "Perform a healthcheck and return a response.", "operationId": "perform_healthcheck_healthcheck_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/img_polluted_object_detection_to_json": {"post": {"summary": "Img Polluted Object Detection To Json", "description": "Object Detection from an image and return detection results in JSON format.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    JSONResponse: JSON with object detection results and pollution status.", "operationId": "img_polluted_object_detection_to_json_img_polluted_object_detection_to_json_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_polluted_object_detection_to_json_img_polluted_object_detection_to_json_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_pretrained_object_detection_to_json": {"post": {"summary": "Img Pretrained Object Detection To Json", "description": "Object Detection from an image and return detection results in JSON format.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    JSONResponse: JSON with object detection results and pollution status.", "operationId": "img_pretrained_object_detection_to_json_img_pretrained_object_detection_to_json_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_pretrained_object_detection_to_json_img_pretrained_object_detection_to_json_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_custom_object_detection_to_json": {"post": {"summary": "Img Object Detection To Json", "description": "Object Detection from an image and return detection results in JSON format.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    JSONResponse: JSON with object detection results and pollution status.", "operationId": "img_object_detection_to_json_img_custom_object_detection_to_json_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_object_detection_to_json_img_custom_object_detection_to_json_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_polluted_object_detection_to_img": {"post": {"summary": "Img Polluted Object Detection To Img", "description": "Object Detection from an image and display annotated images with pollution status.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    StreamingResponse: An image with bbox annotations.", "operationId": "img_polluted_object_detection_to_img_img_polluted_object_detection_to_img_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_polluted_object_detection_to_img_img_polluted_object_detection_to_img_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_pretrained_detection_to_img": {"post": {"summary": "Img Pretrained Detection To Img", "description": "Object Detection from an image and display annotated images with pollution status.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    StreamingResponse: An image with bbox annotations.", "operationId": "img_pretrained_detection_to_img_img_pretrained_detection_to_img_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_pretrained_detection_to_img_img_pretrained_detection_to_img_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/img_custom_detection_to_img": {"post": {"summary": "Img Custom Detection To Img", "description": "Object Detection from an image and display annotated images with pollution status.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    StreamingResponse: An image with bbox annotations.", "operationId": "img_custom_detection_to_img_img_custom_detection_to_img_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_img_custom_detection_to_img_img_custom_detection_to_img_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/get_result": {"post": {"summary": "Get Result", "description": "Get the number of detected objects using all three object detection models.\n\nArgs:\n    file (bytes): The image file in bytes format.\nReturns:\n    JSONResponse: JSON with the number of detected objects for each model.", "operationId": "get_result_get_result_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_get_result_get_result_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/identify_blur_image": {"post": {"summary": "Check Blur", "operationId": "check_blur_identify_blur_image_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_check_blur_identify_blur_image_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_check_blur_identify_blur_image_post": {"title": "Body_check_blur_identify_blur_image_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_get_result_get_result_post": {"title": "Body_get_result_get_result_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_custom_detection_to_img_img_custom_detection_to_img_post": {"title": "Body_img_custom_detection_to_img_img_custom_detection_to_img_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_object_detection_to_json_img_custom_object_detection_to_json_post": {"title": "Body_img_object_detection_to_json_img_custom_object_detection_to_json_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_polluted_object_detection_to_img_img_polluted_object_detection_to_img_post": {"title": "Body_img_polluted_object_detection_to_img_img_polluted_object_detection_to_img_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_polluted_object_detection_to_json_img_polluted_object_detection_to_json_post": {"title": "Body_img_polluted_object_detection_to_json_img_polluted_object_detection_to_json_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_pretrained_detection_to_img_img_pretrained_detection_to_img_post": {"title": "Body_img_pretrained_detection_to_img_img_pretrained_detection_to_img_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "Body_img_pretrained_object_detection_to_json_img_pretrained_object_detection_to_json_post": {"title": "Body_img_pretrained_object_detection_to_json_img_pretrained_object_detection_to_json_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}